Add model validators :: Check User Model and UserController

+++ src/main/java/com/example/demo/models/UserPageModel.java
@@ -12,5 +12,5 @@ public class UserPageModel {
     private List<UserModel> Users;
 
     private int totalPages;
-    private int totalElements;
+    private Long totalElements;
 }
